\name{WinProp-package}
\alias{WinProp-package}
\alias{WinProp}
\docType{package}
\title{
Calculate Win Probabilities and the corresponding confidence intervals
}
\description{
This R package provides the functionality to calculate Win Probabilities and the corresponding confidence intervals. In addition, the confidence intervals for the critical value  with a given chance that treatment one is different from treatment two.
}
\details{
\tabular{ll}{
Package: \tab WinProp\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-10-20\cr
License: \tab GPL\cr
}

}
\author{
Andreas Kitsche

Maintainer: Andreas Kitsche <kitsche@biostat.uni-hannover.de>

}
\references{
A. J. Hayter (2013): Inferences on the difference between future observations for comparing two treatments, Journal of Applied Statistics, 40:4, 887-900

Richard H. Browne (2010) The t-Test p Value and Its Relationship to the Effect Size and P(X>Y), The American Statistician, 64:1, 30-33

Kieser, M.,Friede, T., Gondan, M. Assessment of statistical significance and clinical relevance. Statistics in Medicine (2013) 32(10): 1707-1719
}

\keyword{ package }
\seealso{

}
\examples{

}
